Rename flake8-bandit rules from plugins to checks
