+++ src/flake8_bandit/checks/hardcoded_password_default.rs
