+++ src/flake8_bandit/mod.rs
@@ -1,5 +1,5 @@
+pub mod checks;
 mod helpers;
-pub mod plugins;
 
 #[cfg(test)]
 mod tests {
